+++ WebContent/WEB-INF/admin_view_user.jsp
@@ -99,7 +99,8 @@
 						</fieldset>
 					</form>
 				</div>
-				<div class="tab-pane fade" id="records"></div>
+				<div class="tab-pane fade" id="records">
+				</div>
 			</div>
 		</div>
 	</div>
