+++ WebContent/WEB-INF/trainee_navbar.jsp
@@ -13,7 +13,7 @@
 			<li class="nav-item"><a class="nav-link"
 				href="<c:url value = "/${applicationScope.URL_QUIZZES}" />">See quizzes</a></li>
 			<li class="nav-item"><a class="nav-link"
-				href="<c:url value = "/${applicationScope.URL_RESULTS}" />">Results</a></li>
+				href="<c:url value = "/${applicationScope.URL_RESULTS}" />">See results</a></li>
 		</ul>
 		<form class="form-inline my-2 my-lg-0">
 			<a class="btn btn-secondary my-2 my-sm-0"
