+++ WebContent/WEB-INF/trainee_records.jsp
