+++ src/common/Config.java
@@ -19,6 +19,6 @@ public class Config {
     public static final String URL_VIEW_QUIZ = "viewQuiz";
     public static final String URL_START_QUIZ = "startQuiz";
     public static final String URL_CREATE_THEME = "createTheme";
-    public static final String URL_RESULTS = "results";
-	public static final String URL_VIEW_RESULT = "viewResult";
+	public static final String URL_RESULTS = "records";
+	public static final String URL_VIEW_RESULT = "viewRecords";
 }
